backend/requirements.txt
fastapi==0.79.0
firebase_admin==5.2.0
python-dotenv==0.20.0
strawberry-graphql==0.120.0
uvicorn==0.18.2